add qt library
