+++ .github/workflows/blank.yml
@@ -28,7 +28,10 @@ jobs:
       - uses: actions/checkout@v2
       - name: Install tree
         run: |
-          sudo apt install tree 
+          sudo apt install tree
+      - name: Install qt
+	run: |
+          sudo apt-get install qt5-default 
       - name: Install boost
         run: |
           sudo apt-get install libboost-all-dev
