+++ CMakeLists.txt
@@ -1,21 +1,30 @@
 cmake_minimum_required(VERSION 3.14)
-project(BattleGround LANGUAGES CXX)
+project(BattleGround VERSION 0.1 LANGUAGES CXX)
+
+set(PROJECT_NAME BattleGround)
+set(PROJECT_TESTS ${PROJECT_NAME}_tests)
+set(PROJECT_LIBRARY ${PROJECT_NAME}_library)
 
 set(CMAKE_INCLUDE_CURRENT_DIR ON)
-
+set(CMAKE_AUTOUIC ON)
+set(CMAKE_AUTOMOC ON)
+set(CMAKE_AUTORCC ON)
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
 set(Boost_USE_STATIC_LIBS        ON)  # only find static libs
 set(Boost_USE_DEBUG_LIBS         OFF) # ignore debug libs and
 set(Boost_USE_RELEASE_LIBS       ON)  # only find release libs
 set(Boost_USE_MULTITHREADED      ON)
 set(Boost_USE_STATIC_RUNTIME    OFF)
+
+find_package(QT NAMES Qt6 Qt5 COMPONENTS Core Quick REQUIRED)
+find_package(Qt${QT_VERSION_MAJOR} COMPONENTS Core Quick REQUIRED)
 find_package(Boost 1.66.0)
+
 if(Boost_FOUND)
     include_directories(${Boost_INCLUDE_DIRS})
 endif()
 
-include_directories(source)
-add_subdirectory(source)
+include_directories(project/source)
+add_subdirectory(project)
 add_subdirectory(test)
