+++ project/main.cpp
@@ -1,0 +1,112 @@
+#include <Box2D/Box2D.h>
+#include <QGuiApplication>
+#include <QQmlApplicationEngine>
+#include <iostream>
+
+int main(int argc, char *argv[]) {
+#if QT_VERSION < QT_VERSION_CHECK(6, 0, 0)
+  QCoreApplication::setAttribute(Qt::AA_EnableHighDpiScaling);
+#endif
+
+  b2Vec2 gravity(0.0f, -10.0f);
+  b2World world(gravity);
+
+  b2BodyDef groundBodyDef;
+  groundBodyDef.position.Set(0.0f, -10.0f);
+  b2Body *groundBody = world.CreateBody(&groundBodyDef);
+
+  b2PolygonShape groundBox;
+  groundBox.SetAsBox(50.0f, 10.0f);
+
+  groundBody->CreateFixture(&groundBox, 0.0f);
+
+  b2BodyDef bodyDef;
+  bodyDef.type = b2_dynamicBody;
+  bodyDef.position.Set(0.0f, 4.0f);
+  b2Body *body = world.CreateBody(&bodyDef);
+
+  b2PolygonShape dynamicBox;
+  dynamicBox.SetAsBox(1.0f, 1.0f);
+
+  b2FixtureDef fixtureDef;
+  fixtureDef.shape = &dynamicBox;
+  fixtureDef.density = 1.0f;
+  fixtureDef.friction = 0.3f;
+
+  body->CreateFixture(&fixtureDef);
+
+  float timeStep = 1.0f / 60.0f;
+
+  int32 velocityIterations = 6;
+  int32 positionIterations = 2;
+
+  std::cout.precision(2);
+  for (int32 i = 0; i < 60; ++i) {
+    world.Step(timeStep, velocityIterations, positionIterations);
+    b2Vec2 position = body->GetPosition();
+    float angle = body->GetAngle();
+    std::cout << position.x << " " << position.y << " " << angle << std::endl;
+  }
+
+  QGuiApplication app(argc, argv);
+
+  QQmlApplicationEngine engine;
+  const QUrl url(QStringLiteral("qrc:/main.qml"));
+  QObject::connect(&engine, &QQmlApplicationEngine::objectCreated, &app,
+                   [url](QObject *obj, const QUrl &objUrl) {
+                     if (!obj && url == objUrl) QCoreApplication::exit(-1);
+                   },
+                   Qt::QueuedConnection);
+  engine.load(url);
+
+  return app.exec();
+}
+/*
+#include <stdlib.h>
+#include <iomanip>
+#include <iostream>
+
+int main() {
+  b2Vec2 gravity(0.0f, -10.0f);
+  b2World world(gravity);
+
+  b2BodyDef groundBodyDef;
+  groundBodyDef.position.Set(0.0f, -10.0f);
+  b2Body* groundBody = world.CreateBody(&groundBodyDef);
+
+  b2PolygonShape groundBox;
+  groundBox.SetAsBox(50.0f, 10.0f);
+
+  groundBody->CreateFixture(&groundBox, 0.0f);
+
+  b2BodyDef bodyDef;
+  bodyDef.type = b2_dynamicBody;
+  bodyDef.position.Set(0.0f, 4.0f);
+  b2Body* body = world.CreateBody(&bodyDef);
+
+  b2PolygonShape dynamicBox;
+  dynamicBox.SetAsBox(1.0f, 1.0f);
+
+  b2FixtureDef fixtureDef;
+  fixtureDef.shape = &dynamicBox;
+  fixtureDef.density = 1.0f;
+  fixtureDef.friction = 0.3f;
+
+  body->CreateFixture(&fixtureDef);
+
+  float timeStep = 1.0f / 60.0f;
+
+  int32 velocityIterations = 6;
+  int32 positionIterations = 2;
+
+  std::cout.precision(2);
+  for (int32 i = 0; i < 60; ++i) {
+    world.Step(timeStep, velocityIterations, positionIterations);
+    b2Vec2 position = body->GetPosition();
+    float angle = body->GetAngle();
+    std::cout << position.x << " " << position.y << " " << angle << std::endl;
+  }
+
+  return 0;
+}
+*/
