+++ project/main.qml
@@ -1,0 +1,23 @@
+import QtQuick 2.12
+import QtQuick.Controls 2.5
+
+ApplicationWindow {
+    width: 640
+    height: 480
+    visible: true
+    title: qsTr("Scroll")
+
+    ScrollView {
+        anchors.fill: parent
+
+        ListView {
+            id: listView
+            width: parent.width
+            model: 20
+            delegate: ItemDelegate {
+                text: "Item " + (index + 1)
+                width: listView.width
+            }
+        }
+    }
+}
