+++ project/source/basicunit.cpp
@@ -1,0 +1,6 @@
+#include "basicunit.h"
+
+BasicUnit::BasicUnit()
+{
+
+}
