+++ project/source/basicunit.h
@@ -1,0 +1,11 @@
+#ifndef BASICUNIT_H
+#define BASICUNIT_H
+
+class BasicUnit {
+ public:
+  BasicUnit();
+
+  bool isReady() const { return true; }
+};
+
+#endif  // BASICUNIT_H
